fix(dns): resolve eBPF nil pointer panic and improve DNS traffic generation
- Fix nil pointer dereference in GetStats by providing value parameter to iterator
- Improve eBPF test DNS traffic generation with multiple query methods
- Add getent and host commands for better system resolver coverage
- Add delays to allow eBPF processing time

+++ internal/observers/dns/observer_e2e_test.go
@@ -229,7 +229,7 @@ func TestE2E_RateLimiting(t *testing.T) {
 		eventsGenerated, stats.TotalProblems, elapsed, expectedMax)
 
 	// Stats should reflect rate limiting in action
-	assert.Greater(t, stats.TotalProblems, int64(0))
+	assert.Greater(t, int(stats.TotalProblems), 0)
 }
 
 func TestE2E_RepeatProblemDetection(t *testing.T) {
